Fix the analyzer and bugprone findings that mark real defects
Every site in the issue's first two groups was read and either fixed or,
where it is a false positive, silenced with a named NOLINT and a reason
(191 fixed, 42 suppressed). The real defects:

- free_char() freed an unprototyped NPC's short description twice (through
  GET_NAME) and leaked its keywords.
- Bleeding attack, crippling strike, pressure point, the blinding shield,
  poison touch, the dog's charm, and the Menzoberranzan chokers built their
  affects on uninitialized structs, so garbage flags and bonus types reached
  the victim; the chokers also stacked hitroll and could not be removed.
- The bomb commands passed ACTION_* bitmasks where is_action_available()
  takes atSTANDARD/atMOVE/atSWIFT, so each checked the wrong action.
- Inferno Bomb's message read an integer as a $t string, and do_drive passed
  a direction as act()'s object pointer, which NPC act triggers dereference.
- Unbraced USE_FULL_ROUND_ACTION and USE_MOVE_ACTION made silent trap
  detection and an immortal's apply-poison spend actions.
- Score width 160 lives in a signed byte, read back as -96, and never
  applied.
- main() read the getenv() build id after setenv() could invalidate it.
- Zero-byte allocations for an empty board table and for a drink container
  named only by its liquid; leaks in isname_tok(), load_clans(), zmalloc(),
  and touch of corruption; fopen_restricted() and the rename helpers lost
  errno in close(); device creation indexed its spell list unchecked.

Reviewing next to the flagged sites also found a trailing space stepping
do_homelands(), handle_region_help(), and handle_background_help() past the
argument's terminator, and transform_voice_to_observational() writing
past its heap block once a hint held more than twenty "you " rewrites.
Both are fixed.

Ten regression tests cover the paths a test can reach. The findings
baseline drops from 2,304 to 2,071.

+++ src/character/race.c
@@ -465,7 +465,7 @@ bool display_race_info(struct char_data *ch, const char *racename)
   /* build buffer for ability modifiers */
   for (i = 0; i < NUM_ABILITY_MODS; i++)
   {
-    stat_mod = race_list[race].ability_mods[i];
+    stat_mod = (int)race_list[race].ability_mods[i];
     if (stat_mod != 0)
     {
       found = TRUE;
