+++ src/movement/movement_position.c
@@ -380,7 +380,7 @@ int change_position(struct char_data *ch, int new_position)
   if (!ch)
     return 0;
 
-  int old_position = GET_POS(ch);
+  int old_position = (int)GET_POS(ch);
 
   /* we will put some general checks for having your position changed */
